refactor: rename `AlpakkaFtpClient` to `FtpClient`
